Added profile chelys connection and local storage data sections
